pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.4.0",
    "flask>=3.1.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "twilio>=9.4.4",
]
